Initial fork from hseeberger/scala-sbt
Our intention is to customize to build for our particular application
and use case, so doing a "hard fork" and squashing previous commits
to avoid confusion when viewing project history in GitHub. However,
credit where it is due will be added in our README.

Everything contained up to this commit represents a fork from
hseeberger/scala-sbt@4439df7 and the work of the contributors there.